C: Remove redundant casts
- BACKBUFFER_START_ADDRESS: drop (u16) on 0xF000; the literal already fits
- draw_road_scene_change: drop (u16) from eight u8 table-element reads
  (tbl[i] is u8, promoted to int before | and <<; cast invisible)
- draw_road_scene_change: drop (s16) from HL_clamp_val - DE_road (both
  u16; modular equivalence means s16 result is identical without cast)
- draw_road_scene_change: drop (u8) from sbc & 0xFF (mask already
  constrains value to [0,255]; L_low is u8)
- draw_road_scene_change: drop two inner (u8) casts on fast_counter
  arithmetic (value is at most 60, fits in u8 without truncation)
- draw_road_scene_change: drop (s8) from literal 0x7F (127 fits in s8)
- draw_road_fill_road_pattern: drop (u8) from dr_backbuf_1 >> 8 and
  & 0xFF (results already in [0,255]; D and E are u8)
- draw_road_fill_road_pattern: drop (u8) from u8 arithmetic assigned
  to u8 locals (sm_* variables, D, E, L); replace (u8)(x - 1) with x--
  and (u8)(x - 2) with x -= 2 where applicable
- move_hero_car: drop (s16) from speed < 0 check (speed is int with
  max value 695, well within s16 range)

+++ C/ChaseHQ/ChaseHQ-Internal.h
@@ -26,7 +26,7 @@
 #define BACKBUFFER_ROWBYTES   (BACKBUFFER_WIDTH / 8)
 #define BACKBUFFER_HEIGHT     (128)
 #define BACKBUFFER_LENGTH     (BACKBUFFER_ROWBYTES * BACKBUFFER_HEIGHT)
-#define BACKBUFFER_START_ADDRESS ((u16) 0xF000)
+#define BACKBUFFER_START_ADDRESS 0xF000
 #define BACKBUFFER_END_ADDRESS (BACKBUFFER_START_ADDRESS + BACKBUFFER_LENGTH)
 // draw_char writes rows at stride 256; addresses near the end of the buffer
 // overflow past 0xFFFF in Z80 address space, wrapping into ROM (harmless).
